chore: split endpoints
