+++ packages/backend/src/controllers/v2/ProjectController.ts
@@ -2,14 +2,12 @@ import {
     ApiErrorPayload,
     ApiGetAsyncQueryResults,
     ApiSuccessEmpty,
-    ParameterError,
     QueryExecutionContext,
-    isExecuteAsyncDashboardChartRequest,
-    isExecuteAsyncMetricQueryRequest,
-    isExecuteAsyncSavedChartRequest,
-    isExecuteAsyncUnderlyingDataRequest,
     type ApiExecuteAsyncQueryResults,
-    type ExecuteAsyncQueryRequestParams,
+    type ExecuteAsyncDashboardChartRequestParams,
+    type ExecuteAsyncMetricQueryRequestParams,
+    type ExecuteAsyncSavedChartRequestParams,
+    type ExecuteAsyncUnderlyingDataRequestParams,
     type MetricQuery,
 } from '@lightdash/common';
 import {
@@ -107,107 +105,146 @@ export class V2ProjectController extends BaseController {
     @Hidden()
     @Middlewares([allowApiKeyAuthentication, isAuthenticated])
     @SuccessResponse('200', 'Success')
-    @Post('{projectUuid}/query')
-    @OperationId('executeAsyncQuery')
-    async executeAsyncQuery(
+    @Post('{projectUuid}/query/metric-query')
+    @OperationId('executeAsyncMetricQuery')
+    async executeAsyncMetricQuery(
         @Body()
-        body: ExecuteAsyncQueryRequestParams,
+        body: ExecuteAsyncMetricQueryRequestParams,
+        @Path() projectUuid: string,
+        @Request() req: express.Request,
+    ): Promise<ApiExecuteAsyncQueryResponse> {
+        this.setStatus(200);
+        const context = body.context ?? getContextFromHeader(req);
+
+        const metricQuery: MetricQuery = {
+            exploreName: body.query.exploreName,
+            dimensions: body.query.dimensions,
+            metrics: body.query.metrics,
+            filters: body.query.filters,
+            sorts: body.query.sorts,
+            limit: body.query.limit,
+            tableCalculations: body.query.tableCalculations,
+            additionalMetrics: body.query.additionalMetrics,
+            customDimensions: body.query.customDimensions,
+            timezone: body.query.timezone,
+            metricOverrides: body.query.metricOverrides,
+        };
+
+        const results = await this.services
+            .getProjectService()
+            .executeAsyncMetricQuery({
+                user: req.user!,
+                projectUuid,
+                invalidateCache: body.invalidateCache,
+                metricQuery,
+                context: context ?? QueryExecutionContext.API,
+                granularity: body.query.granularity,
+            });
+
+        return {
+            status: 'ok',
+            results,
+        };
+    }
+
+    @Hidden()
+    @Middlewares([allowApiKeyAuthentication, isAuthenticated])
+    @SuccessResponse('200', 'Success')
+    @Post('{projectUuid}/query/chart')
+    @OperationId('executeAsyncSavedChartQuery')
+    async executeAsyncSavedChartQuery(
+        @Body()
+        body: ExecuteAsyncSavedChartRequestParams,
         @Path() projectUuid: string,
         @Request() req: express.Request,
     ): Promise<ApiExecuteAsyncQueryResponse> {
         this.setStatus(200);
 
         const context = body.context ?? getContextFromHeader(req);
-        const commonArgs = {
-            user: req.user!,
-            projectUuid,
-            invalidateCache: body.invalidateCache,
+
+        const results = await this.services
+            .getProjectService()
+            .executeAsyncSavedChartQuery({
+                user: req.user!,
+                projectUuid,
+                invalidateCache: body.invalidateCache,
+                chartUuid: body.chartUuid,
+                versionUuid: body.versionUuid,
+                context: context ?? QueryExecutionContext.API,
+            });
+
+        return {
+            status: 'ok',
+            results,
         };
+    }
 
-        if (isExecuteAsyncMetricQueryRequest(body)) {
-            const metricQuery: MetricQuery = {
-                exploreName: body.query.exploreName,
-                dimensions: body.query.dimensions,
-                metrics: body.query.metrics,
-                filters: body.query.filters,
-                sorts: body.query.sorts,
-                limit: body.query.limit,
-                tableCalculations: body.query.tableCalculations,
-                additionalMetrics: body.query.additionalMetrics,
-                customDimensions: body.query.customDimensions,
-                timezone: body.query.timezone,
-                metricOverrides: body.query.metricOverrides,
-            };
+    @Hidden()
+    @Middlewares([allowApiKeyAuthentication, isAuthenticated])
+    @SuccessResponse('200', 'Success')
+    @Post('{projectUuid}/query/dashboard-chart')
+    @OperationId('executeAsyncDashboardChartQuery')
+    async executeAsyncDashboardChartQuery(
+        @Body()
+        body: ExecuteAsyncDashboardChartRequestParams,
+        @Path() projectUuid: string,
+        @Request() req: express.Request,
+    ): Promise<ApiExecuteAsyncQueryResponse> {
+        this.setStatus(200);
 
-            const results = await this.services
-                .getProjectService()
-                .executeAsyncMetricQuery({
-                    ...commonArgs,
-                    metricQuery,
-                    context: context ?? QueryExecutionContext.API,
-                    granularity: body.query.granularity,
-                });
+        const context = body.context ?? getContextFromHeader(req);
 
-            return {
-                status: 'ok',
-                results,
-            };
-        }
+        const results = await this.services
+            .getProjectService()
+            .executeAsyncDashboardChartQuery({
+                user: req.user!,
+                projectUuid,
+                invalidateCache: body.invalidateCache,
+                chartUuid: body.chartUuid,
+                dashboardUuid: body.dashboardUuid,
+                dashboardFilters: body.dashboardFilters,
+                dashboardSorts: body.dashboardSorts,
+                granularity: body.granularity,
+                context: context ?? QueryExecutionContext.API,
+            });
 
-        if (isExecuteAsyncSavedChartRequest(body)) {
-            const results = await this.services
-                .getProjectService()
-                .executeAsyncSavedChartQuery({
-                    ...commonArgs,
-                    chartUuid: body.chartUuid,
-                    versionUuid: body.versionUuid,
-                    context: context ?? QueryExecutionContext.API,
-                });
+        return {
+            status: 'ok',
+            results,
+        };
+    }
 
-            return {
-                status: 'ok',
-                results,
-            };
-        }
+    @Hidden()
+    @Middlewares([allowApiKeyAuthentication, isAuthenticated])
+    @SuccessResponse('200', 'Success')
+    @Post('{projectUuid}/query/underlying-data')
+    @OperationId('executeAsyncUnderlyingDataQuery')
+    async executeAsyncUnderlyingDataQuery(
+        @Body()
+        body: ExecuteAsyncUnderlyingDataRequestParams,
+        @Path() projectUuid: string,
+        @Request() req: express.Request,
+    ): Promise<ApiExecuteAsyncQueryResponse> {
+        this.setStatus(200);
 
-        if (isExecuteAsyncDashboardChartRequest(body)) {
-            const results = await this.services
-                .getProjectService()
-                .executeAsyncDashboardChartQuery({
-                    ...commonArgs,
-                    chartUuid: body.chartUuid,
-                    dashboardUuid: body.dashboardUuid,
-                    dashboardFilters: body.dashboardFilters,
-                    dashboardSorts: body.dashboardSorts,
-                    granularity: body.granularity,
-                    context: context ?? QueryExecutionContext.API,
-                });
+        const context = body.context ?? getContextFromHeader(req);
 
-            return {
-                status: 'ok',
-                results,
-            };
-        }
+        const results = await this.services
+            .getProjectService()
+            .executeAsyncUnderlyingDataQuery({
+                user: req.user!,
+                projectUuid,
+                invalidateCache: body.invalidateCache,
+                underlyingDataSourceQueryUuid:
+                    body.underlyingDataSourceQueryUuid,
+                filters: body.filters,
+                underlyingDataItemId: body.underlyingDataItemId,
+                context: context ?? QueryExecutionContext.API,
+            });
 
-        if (isExecuteAsyncUnderlyingDataRequest(body)) {
-            const results = await this.services
-                .getProjectService()
-                .executeAsyncUnderlyingDataQuery({
-                    ...commonArgs,
-                    underlyingDataSourceQueryUuid:
-                        body.underlyingDataSourceQueryUuid,
-                    filters: body.filters,
-                    underlyingDataItemId: body.underlyingDataItemId,
-                    context: context ?? QueryExecutionContext.API,
-                });
-
-            return {
-                status: 'ok',
-                results,
-            };
-        }
-
-        this.setStatus(400);
-        throw new ParameterError('Invalid async query execution request');
+        return {
+            status: 'ok',
+            results,
+        };
     }
 }
